+++ src/Microsoft.Security.Utilities.Core/HighConfidenceMicrosoftSecurityModels/SEC101_101_AzureClientAppLegacyCredential32.cs
@@ -31,6 +31,11 @@ namespace Microsoft.Security.Utilities
                 return null;
             }
 
+            if (DateTime.TryParse(match, out DateTime result))
+            {
+                return null;
+            }
+
             return base.GetMatchIdAndName(match);
         }
 
@@ -50,6 +55,7 @@ namespace Microsoft.Security.Utilities
         public override IEnumerable<string> GenerateFalsePositiveExamples()
         {
             yield return DateTime.UtcNow.ToString("yyyy-MM-ddTHH:mm:ss.fffffffZ");
+            yield return "2024-03-07T02:50:56.464790+00:00";
         }
